backend/requirements.txt
Flask==3.0.2
SQLAlchemy==2.0.25
psycopg[binary,pool]==3.1.18
PyJWT==2.8.0
python-dotenv==1.0.1
waitress==3.0.0
httpx==0.26.0
